Remove header & footer & no-need pages
