+++ src/pages/_app.tsx
@@ -1,12 +1,10 @@
 import '../styles/global.css'
 import 'katex/dist/katex.css'
-import Footer from '../components/footer'
 
 export default function MyApp({ Component, pageProps }) {
   return (
     <>
       <Component {...pageProps} />
-      <Footer />
     </>
   )
 }
